+++ util/logging/logging.go
@@ -92,10 +92,11 @@ func NewLogSet(version semv.Version, name string, err io.Writer) *LogSet {
 	ls := newls(name, WarningLevel, bundle)
 	ls.imposeLevel()
 
+	// use sous.<env>.<region>.*, he said
 	if name == "" {
-		ls.metrics = metrics.NewRegistry()
+		ls.metrics = metrics.NewPrefixedRegistry("sous." + bundle.appIdent.metricsScope())
 	} else {
-		ls.metrics = metrics.NewPrefixedRegistry(name + ".")
+		ls.metrics = metrics.NewPrefixedRegistry(name + "." + bundle.appIdent.metricsScope())
 	}
 	return ls
 }
@@ -107,9 +108,26 @@ func (ls LogSet) Child(name string) LogSink {
 	return child
 }
 
+func getEnvHash() map[string]string {
+	h := map[string]string{}
+	get := func(n string) {
+		if v, has := os.LookupEnv(n); has {
+			h[n] = v
+		}
+	}
+	get("OT_ENV")
+	get("OT_ENV_TYPE")
+	get("OT_ENV_LOCATION")
+	get("TASK_ID")
+	get("INSTANCE_NO")
+	return h
+}
+
 func newdb(vrsn semv.Version, err io.Writer, lgrs *logrus.Logger) *dumpBundle {
+	env := getEnvHash()
+
 	return &dumpBundle{
-		appIdent:   collectAppID(vrsn),
+		appIdent:   collectAppID(vrsn, env),
 		context:    context.Background(),
 		err:        err,
 		defaultErr: err,