+++ CHANGELOG.md
@@ -6,6 +6,13 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/)
 with respect to its command line interface and HTTP interface.
 
+## [Unreleased](//github.com/opentable/sous/compare/0.5.40..HEAD)
+
+### Fixed
+
+* All: ot_* fields are actually populated - also instance-no
+* All: metrics are scoped to env and region so that multiple sous instances don't
+  clobber each other's metrics.
 
 ## [0.5.40](//github.com/opentable/sous/compare/0.5.39..0.5.40)